requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
bson==0.5.10
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2 